Deletions allowed by the previous commit.
[ci skip-rust-tests]
